Fibonacci 98

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
 
 /**
  * main - prints the first 98 Fibonacci numbers, starting with 1 and 2
@@ -9,8 +8,8 @@
 
 int main(void)
 {
-	long float sum;
-	long float first, second;
+	float sum;
+	float first, second;
 	int count;
 
 	sum = 0;
@@ -21,14 +20,14 @@ int main(void)
 	while (count < 97)
 	{
 		sum = first + second;
-		printf("%lf, ", sum);
+		printf("%.0f, ", sum);
 		first = second;
 		second = sum;
 
 		count++;
 	}
 	sum = first + second;
-	printf("%lf\n", sum);
+	printf("%.0f\n", sum);
 
 	return (0);
 }
